refactor(toolbar): extract reusable editor toolbar from onlyoffice
Move the editor toolbar building blocks (BackButton, FileName,
HomeIcon, HomeLinker, Separator, Sharing) out of OnlyOffice into a
shared components/EditorToolbar so other full-screen editors can
reuse them.

FileName now receives isReadOnly as a prop instead of reading it
from the OnlyOffice context, making it independent of OnlyOffice.
